코드 구조 통일
