+++ app/schemas/chat.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+class ChatRequest(BaseModel):
+    user_input: str
+
+class ChatResponse(BaseModel):
+    success: bool
+    status_code: int
+    message: str
